Minor refactoring + restructured namespaces and their references + new test project for API using Moq and xUnit
