+++ PhotoCube/Server/ObjectCubeServer/ObjectCubeServer.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ObjectCubeServer", "ObjectC
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleAppForInteractingWithDatabase", "ConsoleAppForInteractingWithDatabase\ConsoleAppForInteractingWithDatabase.csproj", "{B882F652-8F1E-49AC-909A-B22DB24EC8A9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectDBTests", "ObjectDBTests\ObjectDBTests.csproj", "{6C3522AF-F509-4607-9BB2-9D548BCA8CB8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectCubeServerTests", "ObjectCubeServerTests\ObjectCubeServerTests.csproj", "{6C3522AF-F509-4607-9BB2-9D548BCA8CB8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B279389E-B532-439E-8696-CB955ED8034A}"
 	ProjectSection(SolutionItems) = preProject
