+++ PhotoCube/Server/ObjectCubeServer/ObjectCubeServer/Models/DomainClasses/Node.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace ObjectCubeServer.Models.DomainClasses
 {
