+++ PhotoCube/Server/ObjectCubeServer/ObjectCubeServer/Program.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
+﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
 namespace ObjectCubeServer
@@ -20,7 +14,7 @@ namespace ObjectCubeServer
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
                 .UseStartup<Startup>()
-                .ConfigureLogging((hostingContext, logging) =>
+                .ConfigureLogging((_, logging) =>
                 {
                     logging.AddConsole(); //Logs to console.
                 });
