--- PhotoCube/Server/ObjectCubeServer/ObjectDBTests/UnitTest1.cs
@@ -1,37 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using ObjectCubeServer.Models.DataAccess;
-using ObjectCubeServer.Models.DomainClasses;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace ObjectDBTests
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void CubeObject_Add()
-        {
-            var options = new DbContextOptionsBuilder<ObjectContext>()
-                .UseInMemoryDatabase(databaseName: "TestDb")
-                .Options;
-            
-            using (var context = new ObjectContext(options))
-            {
-                context.CubeObjects.Add(new CubeObject() {
-                    FileType = FileType.Photo,
-                    ObjectTagRelations = new List<ObjectTagRelation>(),
-                });
-                context.SaveChanges();
-            }
-
-            using (var context = new ObjectContext(options))
-            {
-                var allCubeObjects = context.CubeObjects.ToList();
-                Assert.AreEqual(1, allCubeObjects.Count);
-                Assert.AreEqual(42,allCubeObjects.First().Id);
-            }
-        }
-    }
-}
